rename db types for symmetric file structure
